+++ libpod/container_api.go
@@ -187,7 +187,7 @@ func (c *Container) StopWithTimeout(timeout uint) error {
 		return define.ErrCtrStopped
 	}
 
-	return c.stop(timeout)
+	return c.stop(timeout, false)
 }
 
 // Kill sends a signal to a container
@@ -205,12 +205,14 @@ func (c *Container) Kill(signal uint) error {
 		return errors.Wrapf(define.ErrCtrStateInvalid, "can only kill running containers. %s is in state %s", c.ID(), c.state.State.String())
 	}
 
-	defer c.newContainerEvent(events.Kill)
-	if err := c.ociRuntime.killContainer(c, signal); err != nil {
+	// Hardcode all = false, we only use all when removing.
+	if err := c.ociRuntime.KillContainer(c, signal, false); err != nil {
 		return err
 	}
 
 	c.state.StoppedByUser = true
+
+	c.newContainerEvent(events.Kill)
 
 	return c.save()
 }
@@ -221,7 +223,7 @@ func (c *Container) Kill(signal uint) error {
 // Sometimes, the $RUNTIME exec call errors, and if that is the case, the exit code is the exit code of the call.
 // Otherwise, the exit code will be the exit code of the executed call inside of the container.
 // TODO investigate allowing exec without attaching
-func (c *Container) Exec(tty, privileged bool, env, cmd []string, user, workDir string, streams *AttachStreams, preserveFDs int, resize chan remotecommand.TerminalSize, detachKeys string) (int, error) {
+func (c *Container) Exec(tty, privileged bool, env map[string]string, cmd []string, user, workDir string, streams *AttachStreams, preserveFDs uint, resize chan remotecommand.TerminalSize, detachKeys string) (int, error) {
 	var capList []string
 	if !c.batched {
 		c.lock.Lock()
@@ -278,7 +280,19 @@ func (c *Container) Exec(tty, privileged bool, env, cmd []string, user, workDir 
 		user = c.config.User
 	}
 
-	pid, attachChan, err := c.ociRuntime.execContainer(c, cmd, capList, env, tty, workDir, user, sessionID, streams, preserveFDs, resize, detachKeys)
+	opts := new(ExecOptions)
+	opts.Cmd = cmd
+	opts.CapAdd = capList
+	opts.Env = env
+	opts.Terminal = tty
+	opts.Cwd = workDir
+	opts.User = user
+	opts.Streams = streams
+	opts.PreserveFDs = preserveFDs
+	opts.Resize = resize
+	opts.DetachKeys = detachKeys
+
+	pid, attachChan, err := c.ociRuntime.ExecContainer(c, sessionID, opts)
 	if err != nil {
 		ec := define.ExecErrorCodeGeneric
 		// Conmon will pass a non-zero exit code from the runtime as a pid here.
@@ -524,7 +538,10 @@ func (c *Container) WaitWithInterval(waitTimeout time.Duration) (int32, error) {
 		return -1, define.ErrCtrRemoved
 	}
 
-	exitFile := c.exitFilePath()
+	exitFile, err := c.exitFilePath()
+	if err != nil {
+		return -1, err
+	}
 	chWait := make(chan error, 1)
 
 	defer close(chWait)
@@ -639,7 +656,7 @@ func (c *Container) Sync() error {
 		(c.state.State != define.ContainerStateConfigured) &&
 		(c.state.State != define.ContainerStateExited) {
 		oldState := c.state.State
-		if err := c.ociRuntime.updateContainerStatus(c, true); err != nil {
+		if err := c.ociRuntime.UpdateContainerStatus(c, true); err != nil {
 			return err
 		}
 		// Only save back to DB if state changed
@@ -687,7 +704,7 @@ func (c *Container) Refresh(ctx context.Context) error {
 
 	// Next, if the container is running, stop it
 	if c.state.State == define.ContainerStateRunning {
-		if err := c.stop(c.config.StopTimeout); err != nil {
+		if err := c.stop(c.config.StopTimeout, false); err != nil {
 			return err
 		}
 	}
@@ -696,8 +713,10 @@ func (c *Container) Refresh(ctx context.Context) error {
 	if len(c.state.ExecSessions) > 0 {
 		logrus.Infof("Killing %d exec sessions in container %s. They will not be restored after refresh.",
 			len(c.state.ExecSessions), c.ID())
-		if err := c.ociRuntime.execStopContainer(c, c.config.StopTimeout); err != nil {
-			return err
+	}
+	for _, session := range c.state.ExecSessions {
+		if err := c.ociRuntime.ExecStopContainer(c, session.ID, c.StopTimeout()); err != nil {
+			return errors.Wrapf(err, "error stopping exec session %s of container %s", session.ID, c.ID())
 		}
 	}
