+++ libpod/oci_attach_linux.go
@@ -47,7 +47,11 @@ func (c *Container) attach(streams *AttachStreams, keys string, resize <-chan re
 
 	registerResizeFunc(resize, c.bundlePath())
 
-	socketPath := buildSocketPath(c.AttachSocketPath())
+	attachSock, err := c.AttachSocketPath()
+	if err != nil {
+		return err
+	}
+	socketPath := buildSocketPath(attachSock)
 
 	conn, err := net.DialUnix("unixpacket", nil, &net.UnixAddr{Name: socketPath, Net: "unixpacket"})
 	if err != nil {
@@ -108,7 +112,11 @@ func (c *Container) attachToExec(streams *AttachStreams, keys string, resize <-c
 	logrus.Debugf("Attaching to container %s exec session %s", c.ID(), sessionID)
 
 	// set up the socket path, such that it is the correct length and location for exec
-	socketPath := buildSocketPath(c.execAttachSocketPath(sessionID))
+	sockPath, err := c.execAttachSocketPath(sessionID)
+	if err != nil {
+		return err
+	}
+	socketPath := buildSocketPath(sockPath)
 
 	// 2: read from attachFd that the parent process has set up the console socket
 	if _, err := readConmonPipeData(attachFd, ""); err != nil {
